requirements.txt
pandas>=2.0
numpy>=1.26
scikit-learn>=1.3
statsmodels>=0.14
yfinance>=0.2
pyarrow>=14
mlflow>=2.12
dvc>=3.0
pyyaml>=6.0
matplotlib>=3.8
plotly>=5.20
numba>=0.59
pytest>=8.0
python-dotenv>=1.0